requirements.txt
fastapi
uvicorn
selenium
webdriver-manager
python-dotenv
pandas
gspread
oauth2client